장바구니 담긴 상품 삭제, 주문에 담긴 상품 삭제, 상품의 이미지 삭제, 상품 삭제 테스트 확인

+++ boot501/src/main/java/com/busanit501/boot501/shop/service/ItemImgService.java
@@ -11,6 +11,8 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import org.thymeleaf.util.StringUtils;
 
+import java.util.Optional;
+
 @Service
 @RequiredArgsConstructor
 @Transactional
@@ -59,14 +61,20 @@ public class ItemImgService {
     }
 
     public void deleteItemImg(Long itemImgId) throws Exception{
-   //상품 등록시 , 반드시 이미지 등록
-        ItemImg savedItemImg = itemImgRepository.findById(itemImgId)
-                .orElseThrow(EntityNotFoundException::new);
-        //기존 이미지 파일 삭제
-        if(!StringUtils.isEmpty(savedItemImg.getImgName())) {
-            fileService.deleteFile(itemImgLocation+"/"+
-                    savedItemImg.getImgName());
+        //상품 등록시 , 반드시 이미지 등록
+//        ItemImg savedItemImg = itemImgRepository.findById(itemImgId)
+//                .orElseThrow(EntityNotFoundException::new);
+        Optional<ItemImg> result = itemImgRepository.findById(itemImgId);
+        if(result.isPresent()){
+            ItemImg savedItemImg = result.get();
+
+            //기존 이미지 파일 삭제
+            if(!StringUtils.isEmpty(savedItemImg.getImgName())) {
+                fileService.deleteFile(itemImgLocation+"/"+
+                        savedItemImg.getImgName());
+            }
         }
+
         itemImgRepository.deleteById(itemImgId);
     }
